requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
apscheduler
python-dotenv
requests
Flask==2.3.3
Flask-APScheduler==1.13.1
SQLAlchemy==2.0.23
pydantic==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Scraping dependencies
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
apify-client==1.8.0
moviepy==1.0.3
Pillow==10.1.0